Move more files to commons
