+++ functions/api/groups.ts
@@ -1,12 +1,12 @@
 import { parseNewGroupRequest } from '@/common/parse'
+import { errorResponse, jsonResponse, successResponse } from '@/common/response'
+import { getUnixTimestamp } from '@/common/time'
 import type {
   Env,
   NewGroupRequest,
   GroupMetadata,
   GroupValue
 } from '@/common/types'
-import { errorResponse, jsonResponse, successResponse } from '@/response'
-import { getUnixTimestamp } from '@/time'
 
 export const onRequestGet: PagesFunction<Env> = async (context) => {
   console.log('Handling GET request')
