+++ functions/api/ping.ts
@@ -1,4 +1,4 @@
-import { jsonResponse } from '@/response'
+import { jsonResponse } from '@/common/response'
 
 export const onRequestGet: PagesFunction = async () => {
   return jsonResponse({
